changed dicom reader to output float64 again, sobel (others?) get confused

+++ dwi/dicomfile.py
@@ -121,7 +121,7 @@ def construct_image(slices, positions, bvalues):
     """Construct uniform image array from slice dictionary."""
     w, h = slices.values()[0].shape
     shape = (len(positions), w, h, len(bvalues))
-    image = np.empty(shape, dtype=np.float32)
+    image = np.empty(shape, dtype=np.float)
     image.fill(np.nan)
     for k, v in slices.iteritems():
         i = positions.index(k[0])
@@ -147,7 +147,7 @@ def get_bvalue(df):
 
 def get_pixels(df):
     """Return rescaled pixel array from DICOM object."""
-    pixels = df.pixel_array.astype(np.float32)
+    pixels = df.pixel_array.astype(np.float)
     pixels = pixels * df.RescaleSlope + df.RescaleIntercept
     # # Clipping should not be done.
     # lowest = df.WindowCenter - df.WindowWidth/2
